Added tasks policy

+++ app/Http/Controllers/Api/TaskController.php
@@ -28,6 +28,7 @@ class TaskController extends Controller
      */
     public function store(TaskRequest $request)
     {
+        $this->authorize('create', [Task::class, $request->project_id]);
         $task = Task::create($request->all());
         return new TaskResource($task);
     }
@@ -52,6 +53,7 @@ class TaskController extends Controller
      */
     public function update(TaskRequest $request, Task $task)
     {
+        $this->authorize('update', $task);
         $task = $task->update($request->all());
         return new TaskResource($task);
     }
@@ -64,6 +66,7 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('update', $task);
         $task->delete();
         return ['status' => 'ok'];
     }
